Remove BorrowSanitizer.h
