+++ bsan-pass/BorrowSanitizerPass.cpp
@@ -1,5 +1,4 @@
 #include "BorrowSanitizerPass.h"
-#include "BorrowSanitizer.h"
 #include "llvm/Analysis/GlobalsModRef.h"
 #include "llvm/Analysis/StackLifetime.h"
 #include "llvm/Analysis/ValueTracking.h"
@@ -16,36 +15,6 @@
 #include "llvm/Transforms/Utils/ModuleUtils.h"
 
 using namespace llvm;
-
-PreservedAnalyses BorrowSanitizerPass::run(Module &M,
-                                           ModuleAnalysisManager &MAM) {
-  if (checkIfAlreadyInstrumented(M, "nosanitize_borrow"))
-    return PreservedAnalyses::all();
-
-  BorrowSanitizer ModuleSanitizer(M, MAM);
-
-  bool Modified = false;
-
-  auto &FAM = MAM.getResult<FunctionAnalysisManagerModuleProxy>(M).getManager();
-
-  for (Function &F : M) {
-    Modified |= ModuleSanitizer.instrumentFunction(F, FAM);
-  }
-  if (!Modified)
-    return PreservedAnalyses::all();
-
-  Modified |= ModuleSanitizer.instrumentModule(M);
-
-  PreservedAnalyses PA = PreservedAnalyses::none();
-  // We incrementally update the dominator tree throughout
-  // these analysis passes.
-  PA.preserve<DominatorTreeAnalysis>();
-  // GlobalsAA is considered stateless and does not get invalidated unless
-  // explicitly invalidated; PreservedAnalyses::none() is not enough. Sanitizers
-  // make changes that require GlobalsAA to be invalidated.
-  PA.abandon<GlobalsAA>();
-  return PA;
-}
 
 static llvm::PassPluginLibraryInfo getBorrowSanitizerPluginInfo() {
   return {LLVM_PLUGIN_API_VERSION, "BorrowSanitizer", LLVM_VERSION_STRING,
